add validation upload avatar

+++ backend/Online_KinoTeater/Application/Services/AvatarService/AvatarService.cs
@@ -1,4 +1,5 @@
 using Application.Interfaces.AvatarService;
+using Application.Interfaces.ImageProcessor;
 using Application.Interfaces.Storage;
 using Application.Interfaces.UnitOfWork;
 using Domain.Model.Common;
@@ -7,21 +8,29 @@ namespace Application.Services.AvatarService;
 
 public class AvatarService(
     IFileStorageService fileStorageService,
+    IImageProcessor imageProcessor,
     IUnitOfWork unitOfWork): IAvatarService
 {
     public async Task<Result<string>> SetAsync(
         Domain.Model.Entyties.User user, 
-        Stream file, 
-        string contentType, 
-        string originalFileName, 
+        Stream stream, 
         CancellationToken cancellationToken)
     {
+        var processResult = await imageProcessor.ProcessAsync(stream, cancellationToken);
+        if(!processResult.IsSuccess)
+            return Result<string>.Failure(processResult.Error!);
+        var processedImageStream = processResult.Value;
+        
+        var key = BuildAvatarKey(user.Id);
+        
         if (!string.IsNullOrWhiteSpace(user.AvatarPath))
             await fileStorageService.DeleteAsync(user.AvatarPath, cancellationToken);
 
-        var key = BuildAvatarKey(user.Id, originalFileName);
-
-        var uploadedResult = await fileStorageService.UploadAsync(file, key, contentType, cancellationToken);
+        var uploadedResult = await fileStorageService.UploadAsync(
+            processedImageStream!, 
+            key, 
+            "image/jpeg", 
+            cancellationToken);
         if (!uploadedResult.IsSuccess)
             return Result<string>.Failure($"{uploadedResult.Error}");
 
@@ -32,13 +41,5 @@ public class AvatarService(
         return Result<string>.Success(url.Value!);
     }
     
-    private static string BuildAvatarKey(Guid userId, string originalFileName)
-    {
-        var extension = Path.GetExtension(originalFileName);
-        if (string.IsNullOrWhiteSpace(extension))
-            extension =  ".jpg";
-
-        var key = $"avatars/{userId}/avatar{extension}";
-        return key;
-    }
+    private static string BuildAvatarKey(Guid userId) => $"avatars/{userId}/avatar.jpeg";
 }
